backend/requirements.txt
fastapi
pydantic
openai
requests
python-dotenv
uvicorn

#run: uvicorn main:app --reload